Semester 1 GPU Accelerated Data Mining Algorithm Library Algorithm: P-Value GPU Accelerated Feature Selection Engine

+++ processor/Result.h
@@ -5,7 +5,7 @@ struct Result
 {
 	long startTime;
 	long endTime;
-	float* scores;
+	double* scores;
 };
 
 #endif
